replace ugpu by ugpu2
