+++ examples/unstructured/fvm/fvm_nabla_gtir.py
@@ -29,8 +29,8 @@ from gtc.unstructured.gtir import (
 )
 from gtc.unstructured.gtir_to_nir import GtirToNir
 from gtc.unstructured.nir_passes.merge_horizontal_loops import find_and_merge_horizontal_loops
-from gtc.unstructured.nir_to_ugpu2 import NirToUgpu
-from gtc.unstructured.ugpu2_codegen import UgpuCodeGenerator
+from gtc.unstructured.nir_to_ugpu import NirToUgpu
+from gtc.unstructured.ugpu_codegen import UgpuCodeGenerator
 
 
 vertical_loops = []
