+++ tests/gtc/regression/cpp/vertex2edge.py
@@ -30,8 +30,8 @@ from gtc.unstructured.gtir import (
     VerticalLoop,
 )
 from gtc.unstructured.gtir_to_nir import GtirToNir
-from gtc.unstructured.nir_to_ugpu2 import NirToUgpu
-from gtc.unstructured.ugpu2_codegen import UgpuCodeGenerator
+from gtc.unstructured.nir_to_ugpu import NirToUgpu
+from gtc.unstructured.ugpu_codegen import UgpuCodeGenerator
 
 
 field_in = UField(
